Fix show-user for external users
