+++ apiserver/root.go
@@ -5,7 +5,6 @@ package apiserver
 
 import (
 	"reflect"
-	"strings"
 	"sync"
 	"time"
 
@@ -360,114 +359,12 @@ func (r *apiHandler) GetAuthEntity() state.Entity {
 
 // HasPermission returns true if the logged in user can perform <operation> on <target>.
 func (r *apiHandler) HasPermission(operation description.Access, target names.Tag) (bool, error) {
-	return hasPermission(r.state.UserAccess, r.entity.Tag(), operation, target)
+	return common.HasPermission(r.state.UserAccess, r.entity.Tag(), operation, target)
 }
 
 // UserHasPermission returns true if the passed in user can perform <operation> on <target>.
 func (r *apiHandler) UserHasPermission(user names.UserTag, operation description.Access, target names.Tag) (bool, error) {
-	return hasPermission(r.state.UserAccess, user, operation, target)
-}
-
-type userAccessFunc func(names.UserTag, names.Tag) (description.UserAccess, error)
-
-// newControllerUserFromGroup returns a description.UserAccess that serves
-// as a stand-in for a user that has group access but no explicit user
-// access.
-func newControllerUserFromGroup(everyoneAccess description.UserAccess,
-	userTag names.UserTag) description.UserAccess {
-	everyoneAccess.UserTag = userTag
-	everyoneAccess.UserID = strings.ToLower(userTag.Canonical())
-	everyoneAccess.UserName = userTag.Canonical()
-	return everyoneAccess
-}
-
-// maybeUseGroupPermission returns a description.UserAccess updated
-// with the group permissions that apply to it if higher than
-// current.
-// If the passed UserAccess is empty (controller user lacks permissions)
-// but the group is not, a stand-in will be created to hold the group
-// permissions.
-func maybeUseGroupPermission(
-	userGetter userAccessFunc,
-	externalUser description.UserAccess,
-	controllerTag names.ControllerTag,
-	userTag names.UserTag,
-) (description.UserAccess, error) {
-
-	everyoneTag := names.NewUserTag(EveryoneTagName)
-	everyone, err := userGetter(everyoneTag, controllerTag)
-	if errors.IsNotFound(err) {
-		return externalUser, nil
-	}
-	if err != nil {
-		return description.UserAccess{}, errors.Trace(err)
-	}
-	if description.IsEmptyUserAccess(externalUser) &&
-		!description.IsEmptyUserAccess(everyone) {
-		externalUser = newControllerUserFromGroup(everyone, userTag)
-	}
-
-	if everyone.Access.EqualOrGreaterControllerAccessThan(externalUser.Access) {
-		externalUser.Access = everyone.Access
-	}
-	return externalUser, nil
-}
-
-func hasPermission(userGetter userAccessFunc, entity names.Tag,
-	operation description.Access, target names.Tag) (bool, error) {
-	validForKind := false
-
-	switch operation {
-	case description.LoginAccess, description.AddModelAccess, description.SuperuserAccess:
-		validForKind = target.Kind() == names.ControllerTagKind
-	case description.ReadAccess, description.WriteAccess, description.AdminAccess:
-		validForKind = target.Kind() == names.ModelTagKind
-	}
-
-	if !validForKind {
-		return false, nil
-	}
-
-	userTag, ok := entity.(names.UserTag)
-	if !ok {
-		// lets not reveal more than is strictly necessary
-		return false, nil
-	}
-
-	user, err := userGetter(userTag, target)
-	if err != nil && !errors.IsNotFound(err) {
-		return false, errors.Annotatef(err, "while obtaining %s user", target.Kind())
-	}
-	// there is a special case for external users, a group called everyone@external
-	if target.Kind() == names.ControllerTagKind && !userTag.IsLocal() {
-		controllerTag, ok := target.(names.ControllerTag)
-		if !ok {
-			return false, errors.NotValidf("controller tag")
-		}
-
-		// TODO(perrito666) remove the following section about everyone group
-		// when groups are implemented, this accounts only for the lack of a local
-		// ControllerUser when logging in from an external user that has not been granted
-		// permissions on the controller but there are permissions for the special
-		// everyone group.
-		user, err = maybeUseGroupPermission(userGetter, user, controllerTag, userTag)
-		if err != nil {
-			return false, errors.Trace(err)
-		}
-		if description.IsEmptyUserAccess(user) {
-			return false, nil
-		}
-	}
-	// returning this kind of information would be too much information to reveal too.
-	if errors.IsNotFound(err) {
-		return false, nil
-	}
-	modelPermission := user.Access.EqualOrGreaterModelAccessThan(operation) && target.Kind() == names.ModelTagKind
-	controllerPermission := user.Access.EqualOrGreaterControllerAccessThan(operation) && target.Kind() == names.ControllerTagKind
-	if !controllerPermission && !modelPermission {
-		return false, nil
-	}
-	return true, nil
+	return common.HasPermission(r.state.UserAccess, user, operation, target)
 }
 
 // DescribeFacades returns the list of available Facades and their Versions
